% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_hdr}
\alias{get_hdr}
\title{Get DICOM header}
\usage{
get_hdr(path)
}
\arguments{
\item{path}{String. Directory path to a study series (which should contain DICOM files).}
}
\description{
Extract a single representative DICOM header from a given series.
}
\keyword{internal}
