% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_studies.R
\name{import_studies}
\alias{import_studies}
\title{Import DICOM studies}
\usage{
import_studies(accs, dcm_dir, cred_path, air_url)
}
\arguments{
\item{accs}{Character vector. Accession numbers to download.}

\item{dcm_dir}{String. Directory path for output.}

\item{cred_path}{String. File path to air_login.txt (for \code{air_download}).}

\item{air_url}{String. URL for AIR API.}
}
\description{
Download a representative DICOM file from each series in a set of studies.
}
\seealso{
Other main functions: \code{\link{predict_headers}},
  \code{\link{train_model}}
}
\concept{main functions}
