% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{predict_headers}
\alias{predict_headers}
\title{Predict DICOM headers}
\usage{
predict_headers(acc_dir, models, ref)
}
\arguments{
\item{acc_dir}{String. Directory path to unknown study to be classified.}

\item{models}{List. Pretrained model(s).}

\item{ref}{Reference preprocessed training data.}
}
\description{
Function to predict series classification of a bunch of headers
}
\seealso{
Other main functions: 
\code{\link{import_studies}()},
\code{\link{train_model}()}
}
\concept{main functions}
