% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{preprocess_headers}
\alias{preprocess_headers}
\title{Preprocess training DICOM headers}
\usage{
preprocess_headers(tb)
}
\arguments{
\item{tb}{Dataframe with training set header data.}
}
\description{
Preprocess header data from training set.
}
\keyword{internal}
