% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{preprocess_headers_new}
\alias{preprocess_headers_new}
\title{Preprocess NEW DICOM headers}
\usage{
preprocess_headers_new(tb_new, tb_preproc)
}
\arguments{
\item{tb_new}{Dataframe with new/test case header data.}

\item{tb_preproc}{Dataframe with preprocessed training set header data.}
}
\description{
Preprocess header data from new test cases (while respecting variables from original training set)
}
\keyword{internal}
