% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_model}
\alias{train_model}
\title{Train model}
\usage{
train_model(dcm_dir, gt_labels_path, save_path = "model.Rdata",
  n_cv = 5, repeats = 5)
}
\arguments{
\item{dcm_dir}{String. Directory path to training set, organized like:
studies/series/dicoms.}

\item{gt_labels_path}{String. File path to ground truth labels file \code{gt_labels.csv}.}

\item{save_path}{String or NULL. (Optional) File path to save trained model as a \code{.Rdata} file.}

\item{n_cv}{Number of cross validation folds for model training.}

\item{repeats}{Number of repeats for repeated cross validation.}
}
\description{
Train a model to identify modality (e.g axial FLAIR, T1, T1CE) based on
DICOM header data.
}
