% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{predict_headers}
\alias{predict_headers}
\title{Predict DICOM headers}
\usage{
predict_headers(acc_dir, models, tb_preproc)
}
\arguments{
\item{acc_dir}{String. Directory path to unknown study to be classified.}

\item{models}{List. Pretrained model(s).}

\item{tb_preproc}{Preprocessed training data.}
}
\description{
Function to predict series classification of a bunch of headers
}
